+++ tests/test_identify_utils.py
@@ -18,7 +18,12 @@ from itertools import combinations
 
 from cai_causal_graph import CausalGraph, TimeSeriesCausalGraph
 from cai_causal_graph.exceptions import CausalGraphErrors
-from cai_causal_graph.identify_utils import identify_confounders, identify_instruments, identify_mediators
+from cai_causal_graph.identify_utils import (
+    identify_confounders,
+    identify_instruments,
+    identify_markov_boundary,
+    identify_mediators,
+)
 from cai_causal_graph.type_definitions import EDGE_T
 
 
@@ -62,32 +67,39 @@ class TestIdentifyConfounders(unittest.TestCase):
     def test_graph_1(self):
         # compute confounders between source and destination
         confounders = identify_confounders(self.graph_1, node_1='x', node_2='y')
+        self.assertEqual(len(confounders), 1)
         self.assertSetEqual(set(confounders), {'z'})
         # Confirm you can also pass Nodes
         confounders = identify_confounders(
             self.graph_1, node_1=self.graph_1.get_node('x'), node_2=self.graph_1.get_node('y')
         )
+        self.assertEqual(len(confounders), 1)
         self.assertSetEqual(set(confounders), {'z'})
 
     def test_graph_2(self):
         # compute confounders between source and destination
         confounders = identify_confounders(self.graph_2, node_1='x', node_2='y')
+        self.assertEqual(len(confounders), 1)
         self.assertSetEqual(set(confounders), {'u'})
         # Confirm you can pass mix of Node and identifier
         confounders = identify_confounders(self.graph_2, node_1='x', node_2=self.graph_2.get_node('y'))
+        self.assertEqual(len(confounders), 1)
         self.assertSetEqual(set(confounders), {'u'})
 
     def test_graph_3(self):
         # compute confounders between source and destination
         confounders = identify_confounders(self.graph_3, node_1='x', node_2='y')
+        self.assertEqual(len(confounders), 1)
         self.assertSetEqual(set(confounders), {'u'})
         # Confirm you can pass mix of Node and identifier
         confounders = identify_confounders(self.graph_3, node_1=self.graph_3.get_node('x'), node_2='y')
+        self.assertEqual(len(confounders), 1)
         self.assertSetEqual(set(confounders), {'u'})
 
     def test_graph_4(self):
         # compute confounders between source and destination
         confounders = identify_confounders(self.graph_4, node_1='x', node_2='y')
+        self.assertEqual(len(confounders), 2)
         self.assertSetEqual(set(confounders), {'u', 'z'})
 
     def test_non_dag(self):
@@ -133,10 +145,13 @@ class TestIdentifyConfounders(unittest.TestCase):
                 confounders_1b = identify_confounders(cg, node_1=v, node_2=u)
                 confounders_2a = identify_confounders(cg_rev, node_1=u, node_2=v)
                 confounders_2b = identify_confounders(cg_rev, node_1=v, node_2=u)
+                self.assertEqual(len(confounders_1a), len(confounders_1b))
                 self.assertSetEqual(set(confounders_1a), set(confounders_1b))
+                self.assertEqual(len(confounders_2a), len(confounders_2b))
                 self.assertSetEqual(set(confounders_2a), set(confounders_2b))
                 if (u == 'x' and v == 'y') or (u == 'y' and v == 'x'):
                     # If x, y then the answers should match for the graphs with swapped x - y edge
+                    self.assertEqual(len(confounders_1a), len(confounders_2a))
                     self.assertSetEqual(set(confounders_1a), set(confounders_2a))
 
     def test_bigger_graph(self):
@@ -169,8 +184,10 @@ class TestIdentifyConfounders(unittest.TestCase):
             cg.add_edge(f'yc{i - 1}', f'yc{i}')
 
         confounders = identify_confounders(cg, 'x', 'y')
+        self.assertEqual(len(confounders), 1)
         self.assertSetEqual(set(confounders), {'u'})
         confounders = identify_confounders(cg, 'y', 'x')
+        self.assertEqual(len(confounders), 1)
         self.assertSetEqual(set(confounders), {'u'})
 
     def test_time_series_graph(self):
@@ -185,6 +202,7 @@ class TestIdentifyConfounders(unittest.TestCase):
 
         # compute confounders between source and destination
         confounders = identify_confounders(ts_cg, node_1='x', node_2='y')
+        self.assertEqual(len(confounders), 1)
         self.assertSetEqual(set(confounders), {'z'})
 
 
@@ -199,6 +217,7 @@ class TestIdentifyInstruments(unittest.TestCase):
 
         # identify instrumental variables
         instruments = identify_instruments(cg, source='x', destination='y')
+        self.assertEqual(len(instruments), 1)
         self.assertSetEqual(set(instruments), {'z'})
 
     def test_multiple_instruments(self):
@@ -213,6 +232,7 @@ class TestIdentifyInstruments(unittest.TestCase):
 
         # identify instrumental variables
         instruments = identify_instruments(cg, source='x', destination='y')
+        self.assertEqual(len(instruments), 3)
         self.assertSetEqual(set(instruments), {'z_3', 'z_2', 'z_1'})
 
     def test_diamond_and_multiple_confounders(self):
@@ -230,6 +250,7 @@ class TestIdentifyInstruments(unittest.TestCase):
 
         # identify instrumental variables
         instruments = identify_instruments(cg, source='x', destination='y')
+        self.assertEqual(len(instruments), 3)
         self.assertSetEqual(set(instruments), {'z_1', 'z_2', 'z_3'})
 
     def test_no_instruments(self):
@@ -306,6 +327,7 @@ class TestIdentifyInstruments(unittest.TestCase):
 
         # identify instrumental variables
         instruments = identify_instruments(ts_cg, source='x', destination='y')
+        self.assertEqual(len(instruments), 3)
         self.assertSetEqual(set(instruments), {'z', 'z lag(n=1)', 'x lag(n=1)'})
 
     def test_nodelike(self):
@@ -318,6 +340,7 @@ class TestIdentifyInstruments(unittest.TestCase):
 
         # identify instrumental variables
         instruments = identify_instruments(cg, source=cg.get_node('x'), destination=cg.get_node('y'))
+        self.assertEqual(len(instruments), 1)
         self.assertSetEqual(set(instruments), {'z'})
 
     def test_error_cases(self):
@@ -370,6 +393,7 @@ class TestIdentifyMediators(unittest.TestCase):
 
         # identify mediators
         mediators = identify_mediators(cg, source='x', destination='y')
+        self.assertEqual(len(mediators), 1)
         self.assertSetEqual(set(mediators), {'m'})
 
     def test_multiple_mediators(self):
@@ -384,6 +408,7 @@ class TestIdentifyMediators(unittest.TestCase):
 
         # identify mediators
         mediators = identify_mediators(cg, source='x', destination='y')
+        self.assertEqual(len(mediators), 2)
         self.assertSetEqual(set(mediators), {'m1', 'm2'})
 
     def test_multiple_mediators_diamond(self):
@@ -403,6 +428,7 @@ class TestIdentifyMediators(unittest.TestCase):
 
         # identify mediators
         mediators = identify_mediators(cg, source='x', destination='y')
+        self.assertEqual(len(mediators), 2)
         self.assertSetEqual(set(mediators), {'m1', 'm4'})
 
     def test_chain(self):
@@ -413,6 +439,7 @@ class TestIdentifyMediators(unittest.TestCase):
 
         # identify mediators
         mediators = identify_mediators(cg, source='x', destination='y')
+        self.assertEqual(len(mediators), 1)
         self.assertSetEqual(set(mediators), {'m'})
 
     def test_no_mediators(self):
@@ -477,6 +504,7 @@ class TestIdentifyMediators(unittest.TestCase):
 
         # identify mediators
         mediators = identify_mediators(ts_cg, source='x', destination='y')
+        self.assertEqual(len(mediators), 1)
         self.assertSetEqual(set(mediators), {'m'})
 
     def test_nodelike(self):
@@ -490,6 +518,7 @@ class TestIdentifyMediators(unittest.TestCase):
 
         # identify mediators
         mediators = identify_mediators(cg, source=cg.get_node('x'), destination=cg.get_node('y'))
+        self.assertEqual(len(mediators), 1)
         self.assertSetEqual(set(mediators), {'m'})
 
     def test_error_cases(self):
@@ -529,3 +558,165 @@ class TestIdentifyMediators(unittest.TestCase):
         # call identify_mediators on the effect of y on x
         mediators = identify_mediators(cg, source='y', destination='x')
         self.assertEqual(len(mediators), 0)
+
+
+class TestIdentifyMarkovBoundary(unittest.TestCase):
+    def test_simple(self):
+        # Test CausalGraph.
+        cg = CausalGraph()
+        cg.add_edge('u', 'b')
+        cg.add_edge('v', 'c')
+        cg.add_edge('b', 'a')  # 'b' is a parent of 'a'
+        cg.add_edge('c', 'a')  # 'c' is a parent of 'a'
+        cg.add_edge('a', 'd')  # 'd' is a child of 'a'
+        cg.add_edge('a', 'e')  # 'e' is a child of 'a'
+        cg.add_edge('w', 'f')
+        cg.add_edge('f', 'd')  # 'f' is a parent of 'd', which is a child of 'a'
+        cg.add_edge('d', 'x')
+        cg.add_edge('d', 'y')
+        cg.add_edge('g', 'e')  # 'g' is a parent of 'e', which is a child of 'a'
+        cg.add_edge('g', 'z')
+
+        markov_boundary = identify_markov_boundary(cg, node='a')
+        self.assertEqual(len(markov_boundary), 6)
+        self.assertSetEqual(set(markov_boundary), {'b', 'c', 'd', 'e', 'f', 'g'})
+
+        # Test Skeleton.
+        markov_boundary = identify_markov_boundary(cg.skeleton, node='a')
+        self.assertEqual(len(markov_boundary), 4)
+        self.assertSetEqual(set(markov_boundary), {'b', 'c', 'd', 'e'})
+
+        # Test NodeLike.
+        markov_boundary = identify_markov_boundary(cg, node=cg.get_node('a'))
+        self.assertEqual(len(markov_boundary), 6)
+        self.assertSetEqual(set(markov_boundary), {'b', 'c', 'd', 'e', 'f', 'g'})
+
+        markov_boundary = identify_markov_boundary(cg.skeleton, node=cg.skeleton.get_node('a'))
+        self.assertEqual(len(markov_boundary), 4)
+        self.assertSetEqual(set(markov_boundary), {'b', 'c', 'd', 'e'})
+
+    def test_non_dag(self):
+        # Create a mixed graph
+        cg = CausalGraph()
+        cg.add_edge('x', 'm', edge_type=EDGE_T.UNDIRECTED_EDGE)
+        cg.add_edge('m', 'y')
+        cg.add_edge('u', 'x')
+        cg.add_edge('u', 'y')
+
+        with self.assertRaises(TypeError):
+            identify_markov_boundary(cg, node='x')
+
+        # Confirm it works for its skeleton though.
+        mb = identify_markov_boundary(cg.skeleton, node='x')
+        self.assertEqual(len(mb), 2)
+        self.assertSetEqual(set(mb), {'m', 'u'})
+
+        mb = identify_markov_boundary(cg.skeleton, node='m')
+        self.assertEqual(len(mb), 2)
+        self.assertSetEqual(set(mb), {'x', 'y'})
+
+        mb = identify_markov_boundary(cg.skeleton, node='y')
+        self.assertEqual(len(mb), 2)
+        self.assertSetEqual(set(mb), {'m', 'u'})
+
+        mb = identify_markov_boundary(cg.skeleton, node='u')
+        self.assertEqual(len(mb), 2)
+        self.assertSetEqual(set(mb), {'x', 'y'})
+
+    def test_time_series_graph(self):
+        # Test TimeSeriesCausalGraph.
+        cg = TimeSeriesCausalGraph()
+        cg.add_edge('u', 'b')
+        cg.add_edge('v', 'c')
+        cg.add_edge('b', 'a')  # 'b' is a parent of 'a'
+        cg.add_edge('c', 'a')  # 'c' is a parent of 'a'
+        cg.add_edge('a', 'd')  # 'd' is a child of 'a'
+        cg.add_edge('a', 'e')  # 'e' is a child of 'a'
+        cg.add_edge('w', 'f')
+        cg.add_edge('f', 'd')  # 'f' is a parent of 'd', which is a child of 'a'
+        cg.add_edge('d', 'x')
+        cg.add_edge('d', 'y')
+        cg.add_edge('g', 'e')  # 'g' is a parent of 'e', which is a child of 'a'
+        cg.add_edge('g', 'z')
+        # Add auto-regressive terms.
+        cg.add_edge('a lag(n=1)', 'a')
+        cg.add_edge('b lag(n=1)', 'b')
+        cg.add_edge('c lag(n=1)', 'c')
+        cg.add_edge('d lag(n=1)', 'd')
+        cg.add_edge('e lag(n=1)', 'e')
+        cg.add_edge('f lag(n=1)', 'f')
+        cg.add_edge('g lag(n=1)', 'g')
+        cg.add_edge('u lag(n=1)', 'u')
+        cg.add_edge('v lag(n=1)', 'v')
+        cg.add_edge('w lag(n=1)', 'w')
+        cg.add_edge('x lag(n=1)', 'x')
+        cg.add_edge('y lag(n=1)', 'y')
+        cg.add_edge('z lag(n=1)', 'z')
+
+        markov_boundary = identify_markov_boundary(cg, node='a')
+        self.assertEqual(len(markov_boundary), 9)
+        self.assertSetEqual(
+            set(markov_boundary), {'a lag(n=1)', 'b', 'c', 'd', 'd lag(n=1)', 'e', 'e lag(n=1)', 'f', 'g'}
+        )
+
+        # Test Skeleton.
+        markov_boundary = identify_markov_boundary(cg.skeleton, node='a')
+        self.assertEqual(len(markov_boundary), 5)
+        self.assertSetEqual(set(markov_boundary), {'a lag(n=1)', 'b', 'c', 'd', 'e'})
+
+        # Test NodeLike.
+        markov_boundary = identify_markov_boundary(cg, node=cg.get_node('a'))
+        self.assertEqual(len(markov_boundary), 9)
+        self.assertSetEqual(
+            set(markov_boundary), {'a lag(n=1)', 'b', 'c', 'd', 'd lag(n=1)', 'e', 'e lag(n=1)', 'f', 'g'}
+        )
+
+        markov_boundary = identify_markov_boundary(cg.skeleton, node=cg.skeleton.get_node('a'))
+        self.assertEqual(len(markov_boundary), 5)
+        self.assertSetEqual(set(markov_boundary), {'a lag(n=1)', 'b', 'c', 'd', 'e'})
+
+    def test_edge_cases(self):
+        # Floating node so empty MB.
+        cg = CausalGraph()
+        cg.add_edge('x', 'y')
+        cg.add_edge('z', 'x')
+        cg.add_edge('z', 'y')
+        cg.add_node('a')
+
+        markov_boundary = identify_markov_boundary(cg, node='a')
+        self.assertEqual(len(markov_boundary), 0)
+
+        markov_boundary = identify_markov_boundary(cg.skeleton, node='a')
+        self.assertEqual(len(markov_boundary), 0)
+
+        # Let's also test confounding structure.
+        markov_boundary = identify_markov_boundary(cg, node='x')
+        self.assertEqual(len(markov_boundary), 2)
+        self.assertSetEqual(set(markov_boundary), {'z', 'y'})
+
+        markov_boundary = identify_markov_boundary(cg.skeleton, node='x')
+        self.assertEqual(len(markov_boundary), 2)
+        self.assertSetEqual(set(markov_boundary), {'z', 'y'})
+
+        markov_boundary = identify_markov_boundary(cg, node='y')
+        self.assertEqual(len(markov_boundary), 2)
+        self.assertSetEqual(set(markov_boundary), {'z', 'x'})
+
+        markov_boundary = identify_markov_boundary(cg.skeleton, node='y')
+        self.assertEqual(len(markov_boundary), 2)
+        self.assertSetEqual(set(markov_boundary), {'z', 'x'})
+
+        markov_boundary = identify_markov_boundary(cg, node='z')
+        self.assertEqual(len(markov_boundary), 2)
+        self.assertSetEqual(set(markov_boundary), {'x', 'y'})
+
+        markov_boundary = identify_markov_boundary(cg.skeleton, node='z')
+        self.assertEqual(len(markov_boundary), 2)
+        self.assertSetEqual(set(markov_boundary), {'x', 'y'})
+
+        # Finally test error case of node not in graph.
+        with self.assertRaises(CausalGraphErrors.NodeDoesNotExistError):
+            identify_markov_boundary(cg, node='b')
+
+        with self.assertRaises(CausalGraphErrors.NodeDoesNotExistError):
+            identify_markov_boundary(cg.skeleton, node='b')